♻ refactor: KOPIS api XML 명칭과 엔티티 명칭 통일 및 오류 수정
- repository Long 으로 수정
- 엔티티 필드명 KOPIS api XML 명칭과 일치

+++ src/main/java/com/example/sixpark/domain/showtime/repository/ShowTimeRepository.java
@@ -4,5 +4,5 @@ package com.example.sixpark.domain.showtime.repository;
 import com.example.sixpark.domain.showtime.entity.ShowTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowTimeRepository extends JpaRepository<ShowTime, Integer> {
+public interface ShowTimeRepository extends JpaRepository<ShowTime, Long> {
 }
